ACHILLES_HEEL_results.record_count -> BIGINT
Change data type of ACHILLES_HEEL_results.record_count from INT to BIGINT.

+++ inst/sql/sql_server/AchillesHeel_v4.sql
@@ -54,7 +54,7 @@ CREATE TABLE @results_database_schema.ACHILLES_HEEL_results (
   analysis_id INT,
 	ACHILLES_HEEL_warning VARCHAR(255),
 	rule_id INT,
-	record_count INT
+	record_count BIGINT
 );
